Update GH Actions workflow related to deployment
The action helaili/jekyll-action is deprecated
and the repository has been archived. [1]

I described the use of `helaili/jekyll-action` GitHub Actions
in a blog post. [2]

However, I intend to update this post with the current changes.

> This repository has been archived by the owner on Jun 13, 2023.
> It is now read-only.

This is due to changes introduced by GitHub.

> GitHub has a new way to publish a static site
> to Pages using GitHub Actions.
> This means you can now build with the Jekyll command
> (or something else) and use actions to publish,
> all using regular steps.
> This removes the need for Docker based actions
> and gives you all the freedom to configure the execution environment
> (versions of Ruby, extra dependencies...),
> it is a lot faster, easier to debug and closer
> to what you can run on your local machine.

Therefore, the process now takes place in two stages: build & deploy.

The action will be triggered when the default branch is changed
or when invoked manually.

[1]: https://github.com/helaili/jekyll-action
[2]: https://torrocus.com/blog/deployment-jekyll-using-github-actions/